IMDB_DJANGO_BACKEND/requirements.txt
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-filter==2.3.0
django-cors-headers==3.2.1
Pillow==7.1.1
whitenoise==2.0.6